requirements.txt
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
